+++ database/models/data_types/tyre_impression_status.py
@@ -3,7 +3,10 @@ import enum
 
 class TyreImpressionStatus(enum.Enum):
     uploaded = "uploaded"
-    processing = "processing"
-    processed = "processed"
+    queued = "queued"
+    preprocessing = "preprocessing"
+    preprocessed = "preprocessed"
+    extracting_features = "extracting_features"
+    matching= "matching"
     matched = "matched"
     failed = "failed"