+++ tests/helpers/factories/tyre_impression_factory.py
@@ -1,4 +1,4 @@
-import datetime
+from datetime import datetime
 from database.repositories import TyreImpressionRepository
 from database.models.data_types import TyreImpressionStatus
 
@@ -16,7 +16,7 @@ class TyreImpressionFactory:
             "uuid": f"{cls.counter}-uuid",
             "file_path": f"{cls.counter}/file/path",
             "status": TyreImpressionStatus.uploaded,
-            "created_at": datetime.datetime.now(),
+            "created_at": datetime.now(),
         }
 
         defaults.update(kwargs)