+++ tests/models/test_tyre_impression.py
@@ -1,8 +1,8 @@
-from database.models import TyreImpression
+from tests.helpers.factories import TyreImpressionFactory
 
 
 def test_repr():
-    tyre_impression = TyreImpression(file_path="/test/file/path", status="test status")
+    tyre_impression = TyreImpressionFactory.create()
 
     res = tyre_impression.__repr__()
     assert res == f'<TyreImpression {tyre_impression.file_path} {tyre_impression.status}>'