+++ tests/models/test_tyre_model.py
@@ -1,8 +1,8 @@
-from database.models import TyreModel
+from tests.helpers.factories import TyreModelFactory
 
 
 def test_repr():
-    tyre_model = TyreModel(manufacturer="Michelin", model_name="Pilot Sport")
+    tyre_model = TyreModelFactory().create()
 
     res = tyre_model.__repr__()
     assert res == f'<TyreModel {tyre_model.manufacturer} {tyre_model.model_name}>'